requirements.txt
selenium ==4.8.3
beautifulsoup4==4.12.2
redis
AzureOpenAI
numpy
tiktoken
fastapi
uvicorn[standard]
typing
pandas
